refactor: streamline account settings by consolidating components and removing unused settings
