+++ src/ui/src/components/settings/telemetry/index.ts
@@ -1,2 +1,1 @@
 export { TelemetryConsentDialog } from "./TelemetryConsentDialog";
-export { TelemetrySettingsPanel } from "./TelemetrySettingsPanel";
